% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposals.R
\name{add_transmission}
\alias{add_transmission}
\title{Proposal to add a transmission}
\usage{
add_transmission(ctree, pen.prob = 0, pen.size = 1, pen.len = 1)
}
\arguments{
\item{ctree}{Current coloured tree}
}
\description{
Proposal to add a transmission
}
