% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposals.R
\name{add_transmission_4}
\alias{add_transmission_4}
\title{Proposal to add a transmission}
\usage{
add_transmission_4(ctree, bn_weight = 0.8)
}
\arguments{
\item{ctree}{Current coloured tree.}

\item{bn_weight}{Penalisation term for transmitting multiple lineages.}
}
\description{
Proposal to add a transmission
}
