% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposals.R
\name{add_transmission_5}
\alias{add_transmission_5}
\title{Proposal to add a transmission}
\usage{
add_transmission_5(ctree, bn_weight = 1)
}
\arguments{
\item{ctree}{Current coloured tree.}

\item{bn_weight}{Penalisation term for transmitting multiple lineages.}
}
\description{
Proposal to add a transmission
}
