% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_prim_obs}
\alias{calc_prim_obs}
\title{Calculate primary observation times}
\usage{
calc_prim_obs(ptree)
}
\arguments{
\item{ptree}{Phylogenetic tree with host data}
}
\description{
Calculate primary observation times
}
