% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{computeMatWIWM}
\alias{computeMatWIWM}
\title{Build a matrix of probability of who infected whom from a MCMC output}
\usage{
computeMatWIWM(record, burnin = 0.5)
}
\arguments{
\item{record}{Output from inferTTreeM function}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}
}
\value{
Matrix of probability of who infected whom
}
\description{
Build a matrix of probability of who infected whom from a MCMC output
}
