% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{convertToCodaM}
\alias{convertToCodaM}
\title{Convert resTransPhyloM object to coda mcmc format}
\usage{
convertToCodaM(record, burnin = 0.5)
}
\arguments{
\item{record}{Output from inferTTreeM function}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}
}
\value{
Object of class mcmc from coda package
}
\description{
Convert resTransPhyloM object to coda mcmc format
}
