% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractions.R
\name{extractPTreeM}
\alias{extractPTreeM}
\title{Extracts phylogenetic tree from a combined phylogenetic/transmission tree}
\usage{
extractPTreeM(ctree)
}
\arguments{
\item{ctree}{Combined tree}
}
\description{
Extracts phylogenetic tree from a combined phylogenetic/transmission tree
}
