% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractions.R
\name{extractTTreeM}
\alias{extractTTreeM}
\title{Extracts transmission tree from a combined phylogenetic/transmission tree}
\usage{
extractTTreeM(ctree)
}
\arguments{
\item{ctree}{Combined tree}
}
\description{
Extracts transmission tree from a combined phylogenetic/transmission tree
}
