% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inferTTreeM}
\alias{inferTTreeM}
\title{Infer transmission tree given a phylogenetic tree}
\usage{
inferTTreeM(
  ptree,
  w.shape = 2,
  w.scale = 1,
  ws.shape = NA,
  ws.scale = NA,
  w.mean = NA,
  w.std = NA,
  ws.mean = NA,
  ws.std = NA,
  mcmcIterations = 1000,
  thinning = 1,
  tree_updates = 1,
  start_const = 2,
  start_rate = 2,
  startOff.r = 1,
  startOff.p = 0.5,
  startPi = 0.5,
  updateconst = TRUE,
  updaterate = TRUE,
  updateOff.r = TRUE,
  updateOff.p = FALSE,
  updatePi = TRUE,
  qconst = NA,
  qrate = NA,
  qOff.r = NA,
  qOff.p = NA,
  qPi = NA,
  bw = 0.1,
  rd = 1,
  startCTree = NA,
  updateTTree = TRUE,
  dateS = -Inf,
  dateT = Inf,
  delta = 1/365,
  verbose = F
)
}
\arguments{
\item{ptree}{Phylogenetic tree}

\item{w.shape}{Shape parameter of the Gamma distribution representing the generation time}

\item{w.scale}{Scale parameter of the Gamma distribution representing the generation time}

\item{ws.shape}{Shape parameter of the Gamma distribution representing the sampling time}

\item{ws.scale}{Scale parameter of the Gamma distribution representing the sampling time}

\item{w.mean}{Mean of the Gamma distribution representing the generation time}

\item{w.std}{Std of the Gamma distribution representing the generation time}

\item{ws.mean}{Mean of the Gamma distribution representing the sampling time}

\item{ws.std}{Std of the Gamma distribution representing the sampling time}

\item{mcmcIterations}{Number of MCMC iterations to run the algorithm for}

\item{thinning}{MCMC thinning interval between two sampled iterations}

\item{start_const}{Starting value of within-host population size}

\item{start_rate}{Starting value for the within-host population growth rate}

\item{startOff.r}{Starting value of parameter off.r}

\item{startOff.p}{Starting value of parameter off.p}

\item{startPi}{Starting value of sampling proportion pi}

\item{updateconst}{Whether of not to update the parameter lm_const}

\item{updaterate}{Whether of not to update the parameter lm_rate}

\item{updateOff.r}{Whether or not to update the parameter off.r}

\item{updateOff.p}{Whether or not to update the parameter off.p}

\item{updatePi}{Whether or not to update the parameter pi}

\item{qconst}{Proposal kernel range for parameter lm_const}

\item{qrate}{Proposal kernel range for parameter lm_rate}

\item{qOff.r}{Proposal kernel range for parameter Off.r}

\item{qOff.p}{Proposal kernel range for parameter Off.p}

\item{qPi}{Proposal kernel range for parameter pi}

\item{startCTree}{Optional combined tree to start from}

\item{updateTTree}{Whether or not to update the transmission tree}

\item{dateS}{Date when observations start}

\item{dateT}{Date when process stops (this can be Inf for fully simulated outbreaks)}

\item{delta}{Grid precision (smaller is better but slower)}

\item{verbose}{Whether or not to use verbose mode (default is false)}
}
\description{
Infer transmission tree given a phylogenetic tree
}
