% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{log_lik_ptree_given_ctree}
\alias{log_lik_ptree_given_ctree}
\title{Likelihood evaluation for a phylogenetic tree conditional on a transmission tree}
\usage{
log_lik_ptree_given_ctree(ctree, kappa, lambda, hosts = NA)
}
\arguments{
\item{ctree}{Combined tree}

\item{kappa}{Initial pathogen population}

\item{lambda}{Pathogen growth rate}

\item{hosts}{Hosts over which likelihood is calculated}
}
\description{
Likelihood evaluation for a phylogenetic tree conditional on a transmission tree
}
