% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_conversions.R
\name{medTTreeM}
\alias{medTTreeM}
\title{Return the medoid from a resTransPhyloM}
\usage{
medTTreeM(record, burnin = 0.5)
}
\arguments{
\item{record}{Output from inferTTreeM function}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}
}
\value{
The medoid
}
\description{
Return the medoid from a resTransPhyloM
}
