% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ongoing.R
\name{num_approx_disc}
\alias{num_approx_disc}
\title{Numeric approximations (discretised) of branching process}
\usage{
num_approx_disc(
  grid,
  delta,
  off.r,
  off.p,
  pi,
  w.shape,
  w.scale,
  ws.shape,
  ws.scale,
  dateS,
  dateT
)
}
\arguments{
\item{grid}{Discrete grid over which to evaluate functions.}

\item{delta}{Discrete time step.}

\item{off.r}{Shape parameter for the number of offspring.}

\item{off.p}{Probability parameter for the number of offspring.}

\item{pi}{Probability of host being sampled.}

\item{w.shape}{Shape parameter of generation time distribution.}

\item{w.scale}{Scale parameter of generation time distribution.}

\item{ws.shape}{Shape parameter of primary sampling time distribution.}

\item{ws.scale}{Scale parameter of primary sampling time distribution.}

\item{dateS}{Start time of outbreak sampling.}

\item{dateT}{Stop time of outbreak sampling.}
}
\description{
Numeric approximations (discretised) of branching process
}
