% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_conversions.R
\name{phyloFromPTreeM}
\alias{phyloFromPTreeM}
\title{Converts a phylogenetic tree into an ape phylo object}
\usage{
phyloFromPTreeM(ptreem)
}
\arguments{
\item{ptreem}{phylogenetic tree}
}
\value{
phylo object
}
\description{
Converts a phylogenetic tree into an ape phylo object
}
