% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.ptreem}
\alias{plot.ptreem}
\title{Plotting for ptreem}
\usage{
\method{plot}{ptreem}(x, ...)
}
\arguments{
\item{x}{Object of class ptreem, ie  a phylogenetic tree}

\item{...}{Additional parameters are passed on to ape::plot.phylo}
}
\value{
Plot of ptreem
}
\description{
Plotting for ptreem
}
