% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.resTransPhyloM}
\alias{plot.resTransPhyloM}
\title{Plotting for resTransPhylo}
\usage{
\method{plot}{resTransPhyloM}(x, ...)
}
\arguments{
\item{x}{Output from inferTTree}

\item{...}{Additional parameters are passed on}
}
\value{
Plot of TransPhylo results
}
\description{
Plotting for resTransPhylo
}
