% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.ttreem}
\alias{plot.ttreem}
\title{Plotting for ttreem}
\usage{
\method{plot}{ttreem}(x, type = "summarised", w.shape = NA, w.scale = NA, ...)
}
\arguments{
\item{x}{Object of class ttreem, ie  a transmission tree}

\item{type}{Type of plot to display, can be 'detailed' or 'summarised' (default)}

\item{w.shape}{Shape parameter of the generation time, needed for detailed plot only}

\item{w.scale}{Scale parameter of the generation time, needed for detailed plot only}

\item{...}{Additional parameters are passed on}
}
\description{
Plotting for ttreem
}
