% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTracesM}
\alias{plotTracesM}
\title{Plot MCMC traces}
\usage{
plotTracesM(record, burnin = 0)
}
\arguments{
\item{record}{Output from inferTTree function}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}

\item{extend}{Whether to also show traces of off.r and off.p}
}
\value{
Returns invisibly the first parameter
}
\description{
Plot MCMC traces
}
