% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTTree_summ}
\alias{plotTTree_summ}
\title{Plot a transmission tree in an economic format}
\usage{
plotTTree_summ(ttree, showLabels = T, maxTime = NA, cex = 1)
}
\arguments{
\item{ttree}{Transmission tree}

\item{showLabels}{Boolean for whether or not to show the labels}

\item{maxTime}{Maximum value of time to show on x axis}

\item{cex}{Expansion factor}
}
\description{
Plot a transmission tree in an economic format
}
