% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ptreem}
\alias{print.ptreem}
\title{Print function for ptreem objects}
\usage{
\method{print}{ptreem}(x, ...)
}
\arguments{
\item{x}{Object of class ptreem, ie a phylogenetic tree}

\item{...}{Additional parameters are passed on}
}
\description{
Print function for ptreem objects
}
