% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{print.resTransPhyloM}
\alias{print.resTransPhyloM}
\title{Print function for resTransPhyloM objects}
\usage{
\method{print}{resTransPhyloM}(x, ...)
}
\arguments{
\item{x}{output from inferTTreeM}

\item{...}{Additional parameters are passed on}
}
\value{
Print out details of TransPhyloMulti results
}
\description{
Print function for resTransPhyloM objects
}
