% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ttreem}
\alias{print.ttreem}
\title{Print function for ttreem objects}
\usage{
\method{print}{ttreem}(x, ...)
}
\arguments{
\item{x}{Object of class ttreem, ie a transmission tree}

\item{...}{Additional parameters are passed on}
}
\description{
Print function for ttreem objects
}
