% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_conversions.R
\name{ptreeFromPhyloM}
\alias{ptreeFromPhyloM}
\title{Converts an ape phylo object into a phylogenetic tree with host information}
\usage{
ptreeFromPhyloM(tr, dateLastSample, host)
}
\arguments{
\item{tr}{phylo object}

\item{dateLastSample}{date of the last sample}

\item{host}{Optional host information, if not present the leaf names needs to contain this, eg 1.1 etc}
}
\value{
phylogenetic tree
}
\description{
Converts an ape phylo object into a phylogenetic tree with host information
}
