% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposals.R
\name{remove_add_4}
\alias{remove_add_4}
\title{Proposal to move a transmission maintaining the same host}
\usage{
remove_add_4(ctree, bn_weight = 0.8, delta = 1)
}
\arguments{
\item{ctree}{Current coloured tree.}

\item{bn_weight}{Penalisation term for transmitting multiple lineages.}

\item{delta}{Proposal magnitude when moving the root transmission.}
}
\description{
Proposal to move a transmission maintaining the same host
}
