% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposals.R
\name{remove_transmission}
\alias{remove_transmission}
\title{Proposal to remove a transmission}
\usage{
remove_transmission(ctree, pen.prob = 0, pen.size = 1, pen.len = 1)
}
\arguments{
\item{ctree}{Current coloured tree.}
}
\description{
Proposal to remove a transmission
}
