% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposals.R
\name{remove_transmission_3}
\alias{remove_transmission_3}
\title{Proposal to remove a transmission}
\usage{
remove_transmission_3(ctree, bn_weight = 0.1)
}
\arguments{
\item{ctree}{Current coloured tree.}

\item{bn_weight}{Penalisation term for transmitting multiple lineages.}
}
\description{
Proposal to remove a transmission
}
