% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_conversions.R
\name{removeMulti}
\alias{removeMulti}
\title{Converts a resTransPhyloM into a resTransPhylo}
\usage{
removeMulti(r)
}
\arguments{
\item{r}{object of class resTransPhyloM which is the output of inferTTreeM}
}
\value{
object of class resTransPhylo produced by removing non-first samples
}
\description{
Converts a resTransPhyloM into a resTransPhylo
}
