% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ongoing.R
\name{sim_ongoing_lim_t}
\alias{sim_ongoing_lim_t}
\title{Simulate an ongoing outbreak with specified observation end date}
\usage{
sim_ongoing_lim_t(
  off.r = 2,
  off.p = 0.5,
  kappa = 0.2,
  lambda = 0.2,
  pi = 0.8,
  sec.p = 0.1,
  sec.t = 14/365,
  w.shape = 2,
  w.scale = 1,
  ws.shape = NULL,
  ws.scale = NULL,
  w.mean = NULL,
  w.std = NULL,
  ws.mean = NULL,
  ws.std = NULL,
  outbreak.start = 2000,
  obs.start = NULL,
  obs.end = 2010,
  grid.delta = 1/365
)
}
\arguments{
\item{off.r}{Shape parameter for the number of offspring}

\item{off.p}{Probability parameter for the number of offspring}

\item{kappa}{Initial pathogen population}

\item{lambda}{Pathogen population linear growth rate}

\item{pi}{Probability of host being sampled}

\item{sec.p}{Probability of repeated samples}

\item{sec.t}{Time between repeated samples}

\item{w.shape}{Shape parameter of generation time distribution}

\item{w.scale}{Scale parameter of generation time distribution}

\item{ws.shape}{Shape parameter of primary sampling time distribution}

\item{ws.scale}{Scale parameter of primary sampling time distribution}

\item{w.mean}{Mean parameter of generation time distribution}

\item{w.std}{Standard deviation of generation time distribution}

\item{ws.mean}{Mean parameter of primary sampling time distribution}

\item{ws.std}{Standard deviation of primary sampling time distribution}

\item{outbreak.start}{Outbreak start date}

\item{obs.start}{Start time of outbreak sampling}

\item{obs.end}{Stop time of outbreak sampling}

\item{grid.delta}{Discrete time step}
}
\description{
Simulate an ongoing outbreak with specified observation end date
}
