% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ongoing.R
\name{sim_ongoing_n}
\alias{sim_ongoing_n}
\title{Simulate an ongoing outbreak}
\usage{
sim_ongoing_n(
  off.r = 1,
  off.p = 0.5,
  lm_const = 0.2,
  lm_rate = 0.2,
  pi = 0.5,
  add.prob = 0.1,
  add.sep = 14/365,
  w.shape = 2,
  w.scale = 1,
  ws.shape = NULL,
  ws.scale = NULL,
  w.mean = NULL,
  w.std = NULL,
  ws.mean = NULL,
  ws.std = NULL,
  dateStartOutbreak = 2000,
  dateS = NULL,
  delta = 1/365,
  n_lim = 200
)
}
\arguments{
\item{off.r}{Shape parameter for the number of offspring.}

\item{off.p}{Probability parameter for the number of offspring.}

\item{lm_const}{Initial pathogen population.}

\item{lm_rate}{Pathogen population linear growth rate.}

\item{pi}{Probability of host being sampled.}

\item{add.prob}{Probability of repeated samples.}

\item{add.sep}{Time between repeated samples.}

\item{w.shape}{Shape parameter of generation time distribution.}

\item{w.scale}{Scale parameter of generation time distribution.}

\item{ws.shape}{Shape parameter of primary sampling time distribution.}

\item{ws.scale}{Scale parameter of primary sampling time distribution.}

\item{w.mean}{Mean parameter of generation time distribution.}

\item{w.std}{Standard deviation of generation time distribution.}

\item{ws.mean}{Mean parameter of primary sampling time distribution.}

\item{ws.std}{Standard deviation of primary sampling time distribution.}

\item{dateStartOutbreak}{Outbreak start date.}

\item{dateS}{Start time of outbreak sampling.}

\item{delta}{Discrete time step.}

\item{n_lim}{Maximum number of samples.}
}
\description{
Simulate an ongoing outbreak
}
