% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{TransPhyloMulti}
\alias{TransPhyloMulti}
\title{TransPhyloMulti: ransPhylo With Multiple Samples Per Host}
\description{
TransPhyloMulti is a R package extending TransPhylo to allow multiple samples per host.
}
\author{
Jake Carson \email{jake.carson@warwick.ac.uk}
}
