% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim_root}
\alias{trim_root}
\title{Remove extra root hosts from a coloured tree}
\usage{
trim_root(ctree)
}
\arguments{
\item{ctree}{Coloured tree}
}
\description{
Remove extra root hosts from a coloured tree
}
