% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trunc_ctree}
\alias{trunc_ctree}
\title{Truncate a combined tree}
\usage{
trunc_ctree(ctree, trunc_time = NA, n_obs = NA)
}
\arguments{
\item{ctree}{Coloured tree}

\item{trunc_time}{Time to truncate tree}

\item{n_obs}{Number of observations in final tree}
}
\description{
Truncate a combined tree
}
