% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation.R
\name{init_ctree}
\alias{init_ctree}
\title{Create a transmission tree compatible with the provided phylogenetic tree}
\usage{
init_ctree(ptree, bottleneck = 1)
}
\arguments{
\item{ptree}{Phylogenetic tree}
}
\description{
Create a transmission tree compatible with the provided phylogenetic tree
}
