% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{log_lik_ptree_given_ctree}
\alias{log_lik_ptree_given_ctree}
\title{Likelihood evaluation for a phylogenetic tree conditional on a transmission tree}
\usage{
log_lik_ptree_given_ctree(ctree, lm_const, lm_rate)
}
\arguments{
\item{ctree}{Combined tree.}

\item{lm_const}{Initial pathogen population.}

\item{lm_rate}{Pathogen growth rate}
}
\description{
Likelihood evaluation for a phylogenetic tree conditional on a transmission tree
}
