% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{log_lik_ttree}
\alias{log_lik_ttree}
\title{Calculate transmission tree likelihood}
\usage{
log_lik_ttree(
  ttree,
  grid,
  fn_list,
  off.r,
  off.p,
  pi,
  w.shape,
  w.scale,
  ws.shape,
  ws.scale,
  dateS,
  dateT,
  delta = 1/365,
  hosts = NA
)
}
\arguments{
\item{ttree}{Transmission tree.}

\item{off.r}{Shape parameter for the number of offspring.}

\item{off.p}{Probability parameter for the number of offspring.}

\item{pi}{Probability of host being sampled.}

\item{w.shape}{Shape parameter of generation time distribution.}

\item{w.scale}{Scale parameter of generation time distribution.}

\item{ws.shape}{Shape parameter of primary sampling time distribution.}

\item{ws.scale}{Scale parameter of primary sampling time distribution.}

\item{dateS}{Start time of outbreak sampling.}

\item{dateT}{Stop time of outbreak sampling.}

\item{delta}{Discrete time step.}
}
\description{
Calculate transmission tree likelihood
}
