% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_conversions.R
\name{ptreeFromPhyloM}
\alias{ptreeFromPhyloM}
\title{Converts an ape phylo object into a phylogenetic tree with host information}
\usage{
ptreeFromPhyloM(tr, dateLastSample)
}
\arguments{
\item{tr}{phylo object with labels indicating host, eg 1.1 etc}

\item{dateLastSample}{date of the last sample}
}
\value{
phylogenetic tree
}
\description{
Converts an ape phylo object into a phylogenetic tree with host information
}
