% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposals.R
\name{remove_add}
\alias{remove_add}
\title{Proposal to move a transmission maintaining the same host}
\usage{
remove_add(ctree, pen.prob = 0, pen.size = 1, pen.len = 1, epsilon = 1)
}
\arguments{
\item{ctree}{Current coloured tree.}
}
\description{
Proposal to move a transmission maintaining the same host
}
